channel validations

+++ RESTfulAPI/Gateways/ChannelMessagesGateway.cs
@@ -1,26 +1,35 @@
 ﻿using Models;
 using RESTfulAPI.Repositories;
+using RESTfulAPI.Repositories.MongoDB;
 
 namespace RESTfulAPI.Gateways;
 
 public class ChannelMessagesGateway : IChannelMessagesGateway
 {
+    private const string UserNotFollowingError = "The user does not follow the specified server";
+
     private ILogger<ChannelMessagesGateway> Logger { get; }
     private IChannelMessagesRepository MessagesRepository { get; }
-    private IServersRepository ServersRepository { get; }
+    private IChannelsRepository ChannelsRepository { get; }
+    private IServerFollowersRepository ServerFollowersRepository { get; }
 
     public ChannelMessagesGateway(
         ILogger<ChannelMessagesGateway> logger,
         IChannelMessagesRepository messagesRepository,
-        IServersRepository serversRepository)
+        IChannelsRepository channelsRepository,
+        IServerFollowersRepository serverFollowersRepository)
     {
         Logger = logger;
         MessagesRepository = messagesRepository;
-        ServersRepository = serversRepository;
+        ChannelsRepository = channelsRepository;
+        ServerFollowersRepository = serverFollowersRepository;
     }
 
-    public async Task<string> Add(string senderId, string channelId, string content, string image)
+    public async Task<Result<string>> Add(string senderId, string channelId, string content, string image)
     {
+        var operationAllowed = await GetIsUserFollowingServer(senderId, channelId);
+        if (!operationAllowed) return Result<string>.Failed(UserNotFollowingError);
+
         var channelMessage = new ChannelMessage
         {
             ChannelId = channelId,
@@ -33,35 +42,61 @@ public class ChannelMessagesGateway : IChannelMessagesGateway
         return await MessagesRepository.Add(channelMessage);
     }
 
-    public async Task<ChannelMessage> Get(string id)
+    public async Task<Result<ChannelMessage>> Get(string senderId, string messageId)
     {
-        return await MessagesRepository.GetById(id);
+        var message = await MessagesRepository.GetById(messageId);
+
+        var operationAllowed = await GetIsUserFollowingServer(senderId, message.ChannelId);
+        if (!operationAllowed) return Result<ChannelMessage>.Failed(UserNotFollowingError);
+
+        return message;
     }
 
-    public async Task<IEnumerable<ChannelMessage>> Get(string channelId, int take)
+    public async Task<Result<IEnumerable<ChannelMessage>>> Get(string senderId, string channelId, int take)
     {
-        return await MessagesRepository.Get(channelId, take, 0);
+        var operationAllowed = await GetIsUserFollowingServer(senderId, channelId);
+        if (!operationAllowed) return Result<IEnumerable<ChannelMessage>>.Failed(UserNotFollowingError);
+
+        return Result<IEnumerable<ChannelMessage>>.Successful(await MessagesRepository.Get(channelId, take, 0));
     }
 
-    public async Task<IEnumerable<ChannelMessage>> GetBefore(DateTime dateTime, string channelId, int take)
+    public async Task<Result<IEnumerable<ChannelMessage>>> GetBefore(string senderId, string channelId, DateTime dateTime, int take)
     {
-        return await MessagesRepository.GetBefore(dateTime, channelId, take);
+        var operationAllowed = await GetIsUserFollowingServer(senderId, channelId);
+        if (!operationAllowed) return Result<IEnumerable<ChannelMessage>>.Failed(UserNotFollowingError);
+
+        return Result<IEnumerable<ChannelMessage>>.Successful(await MessagesRepository.GetBefore(dateTime, channelId, take));
     }
 
-    public async Task<bool> Update(string senderId, string messageId, string newContent)
+    public async Task<Result<bool>> Update(string senderId, string messageId, string newContent)
     {
         if (string.IsNullOrEmpty(newContent)) return false;
-        var message = await Get(messageId);
+
+        var messageResult = await Get(senderId, messageId);
+        if (!messageResult.Success) return Result<bool>.Failed(messageResult.Errors);
+
+        var message = messageResult.Value;
         if (!senderId.Equals(message.SenderId)) return false;
+
         message.Content = newContent;
         message.EditedTimestamp = DateTime.UtcNow;
         return await MessagesRepository.Edit(messageId, message);
     }
 
-    public async Task<bool> Delete(string senderId, string messageId)
+    public async Task<Result<bool>> Delete(string senderId, string messageId)
     {
-        var message = await Get(messageId);
+        var messageResult = await Get(senderId, messageId);
+        if (!messageResult.Success) return Result<bool>.Failed(messageResult.Errors);
+
+        var message = messageResult.Value;
         if (!senderId.Equals(message.SenderId)) return false;
         return await MessagesRepository.Delete(messageId);
+    }
+
+    private async Task<bool> GetIsUserFollowingServer(string senderId, string channelId)
+    {
+        var channelDetails = await ChannelsRepository.GetById(channelId);
+        if (channelDetails == null) return false;
+        return await ServerFollowersRepository.GetIsUserFollowingServer(senderId, channelDetails.ServerId);
     }
 }
